// forms/client-forms/src/api/generated/model/formQuestionAnswer.ts
/**
 * Generated by orval v7.10.0 🍺
 * Do not edit manually.
 * Forms-gateway
 * The gateway API for forms application
 * OpenAPI spec version: 1.0
 */
import type { FormQuestionAnswerValue } from './formQuestionAnswerValue';
import type { EvaluationFormTemplate } from './evaluationFormTemplate';

export interface FormQuestionAnswer {
  id: number;
  formQuestionId: number;
  object: FormQuestionAnswerValue;
  formId: number;
  evaluationFormTemplate?: EvaluationFormTemplate;
}